Updated start directory
