+++ start/gateway/src/main/java/io/openliberty/guides/gateway/InventoryGatewayResource.java
@@ -12,11 +12,11 @@
 // end::copyright[]
 package io.openliberty.guides.gateway;
 
-import java.util.concurrent.CompletionStage;
+import java.util.Properties;
 
+import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
 import javax.ws.rs.GET;
-import javax.ws.rs.POST;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -24,39 +24,30 @@ import javax.ws.rs.core.MediaType;
 
 import org.eclipse.microprofile.rest.client.inject.RestClient;
 
-import io.openliberty.guides.gateway.client.JobClient;
-import io.openliberty.guides.models.JobList;
-import io.openliberty.guides.models.Job;
-import io.openliberty.guides.models.JobResult;
+import io.openliberty.guides.gateway.client.InventoryGatewayClient;
+import io.openliberty.guides.models.InventoryList;
+import io.openliberty.guides.models.SystemData;
 
-@Path("/jobs")
-public class JobResource {
+@RequestScoped
+@Path("/systems")
+public class InventoryGatewayResource {
 
     @Inject
     @RestClient
-    private JobClient jobClient;
+    private InventoryGatewayClient inventoryClient;
 
     @GET
     @Produces(MediaType.APPLICATION_JSON)
-    public JobList getJobs() {
-        try {
-            return new JobList(jobClient.getJobs().getResults());
-        } catch (Exception ex) {
-            // Respond with empty list on error
-            return new JobList();
-        }
+    public InventoryList getSystems() {
+        return inventoryClient.getInventory();
     }
 
     @GET
-    @Path("{jobId}")
+    @Path("{hostname}")
     @Produces(MediaType.APPLICATION_JSON)
-    public JobResult getJob(@PathParam("jobId") String jobId) {
-        return jobClient.getJob(jobId);
+    public SystemData getSystem(@PathParam("hostname") String hostname) {
+        Properties properties = inventoryClient.getProperties(hostname);
+        return new SystemData(hostname, properties);
     }
 
-    @POST
-    @Produces(MediaType.APPLICATION_JSON)
-    public Job createJob() {
-        return jobClient.createJob();
-    }
 }
